+++ R/ggMarginal.R
@@ -38,9 +38,12 @@
 #' while \code{ggMarginal(p, xparams = list(size=2), yparams = list(size=2))}
 #' will make the density plot outline thicker.
 #' @examples
+#' 
+#' # basic usage
 #' p <- ggplot2::ggplot(mtcars, ggplot2::aes(wt, mpg)) + ggplot2::geom_point()
 #' ggMarginal(p)
 #'       
+#' # using some parameters
 #' set.seed(30)
 #' df <- data.frame(x = rnorm(500, 50, 10), y = runif(500, 0, 50))
 #' p2 <- ggplot2::ggplot(df, ggplot2::aes(x, y)) + ggplot2::geom_point()
@@ -50,18 +53,26 @@
 #' ggMarginal(p2, size = 2)
 #' ggMarginal(p2, colour = "red")
 #' ggMarginal(p2, colour = "red", xparams = list(colour = "blue", size = 3))
-#' p2 <- p2 + ggplot2::ggtitle("Random data") + ggplot2::theme_bw(30)
-#' ggMarginal(p2)
-#'       
+#' ggMarginal(p2, type = "histogram", bins = 10)
+#' 
+#' # specifying the data directly instead of providing a plot       
 #' ggMarginal(data = df, x = "x", y = "y")
-#'       
+#'
+#' # more examples showing how the marginal plots are properly aligned even when
+#' # the main plot axis/margins/size/etc are changed
 #' set.seed(30)
 #' df2 <- data.frame(x = c(rnorm(250, 50, 10), rnorm(250, 100, 10)),
 #'                   y = runif(500, 0, 50))
-#' p3 <- ggplot2::ggplot(df2, ggplot2::aes(x, y)) + ggplot2::geom_point()
+#' p2 <- ggplot2::ggplot(df2, ggplot2::aes(x, y)) + ggplot2::geom_point()
+#' ggMarginal(p2)
+#' 
+#' p2 <- p2 + ggplot2::ggtitle("Random data") + ggplot2::theme_bw(30)
+#' ggMarginal(p2)
+#' 
+#' p3 <- ggplot2::ggplot(df2, ggplot2::aes(log(x), y - 500)) + ggplot2::geom_point()
 #' ggMarginal(p3)
 #' 
-#' p4 <- ggplot2::ggplot(df2, ggplot2::aes(log(x), y - 500)) + ggplot2::geom_point()
+#' p4 <- p3 + ggplot2::scale_x_continuous(limits = c(2, 6)) + ggplot2::theme_bw(50)
 #' ggMarginal(p4)
 #' @seealso \href{http://daattali.com/shiny/ggExtra-ggMarginal-demo/}{Demo Shiny app}
 #' @export
@@ -136,14 +147,12 @@ ggMarginal <- function(p, data, x, y, type = c("density", "histogram", "boxplot"
   
   # Remove all margin around plot so that it's easier to position the
   # density plots beside the main plot
-  p <- p + ggplot2::theme(plot.margin = grid::unit(c(0, 0, 0, 0), "mm"))
+  p <- p + ggplot2::theme(plot.margin = grid::unit(c(0, 0, 0, 0), "null"))
 
   # Decompose the original ggplot2 object to grab all sorts of information from it
   pb <- ggplot2::ggplot_build(p)
-  
-  textsize <- p$theme$text$size
 
-  # get the common code for both maginal (x and y) plots
+  # get the common code for both marginal (x and y) plots
   marginPlot <- function(margin) {
     aest <- ggplot2::aes()
     if (margin == "x") {
@@ -193,10 +202,108 @@ ggMarginal <- function(p, data, x, y, type = c("density", "histogram", "boxplot"
     plot + layer
   }  
   
+  # given a plot, copy some theme properties from the main plot so that they will
+  # resemble each other more and look better beside each other, and also add
+  # some common theme properties such as 0 margins and transparent text colour
+  addMainTheme <- function(marginal, margin) {
+    if (packageVersion("ggplot2") > "1.0.1") {
+      # copy theme from main plot
+      themeProps <- c("text",
+                      "axis.text","axis.text.x", "axis.text.y",
+                      "axis.ticks", "axis.ticks.length",
+                      "axis.title", "axis.title.x", "axis.title.y",
+                      "plot.title"
+                    )
+      for(property in themeProps) {
+        marginal$theme[[property]] <- p$theme[[property]]
+      }
+      
+      # make text and line colours transparent
+      transparentProps <- c("text",
+                            "axis.text", "axis.text.x", "axis.text.y",
+                            "axis.ticks",
+                            "axis.title", "axis.title.x", "axis.title.y",
+                            "line")
+      for(property in transparentProps) {
+        if (!is.null(marginal$theme[[property]])) {
+          marginal$theme[[property]]$colour <- "transparent"
+        } else if (property %in% c("axis.ticks", "line")) {
+          themePair <- list()
+          themePair[[property]] <- ggplot2::element_line(colour = "transparent")
+          marginal <- marginal + do.call(ggplot2::theme, themePair)
+        } else {
+          themePair <- list()
+          themePair[[property]] <- ggplot2::element_text(colour = "transparent")
+          marginal <- marginal + do.call(ggplot2::theme, themePair)
+        }
+      }
+      
+      # some more theme properties
+      marginal <- marginal +
+        ggplot2::theme(
+          panel.background = ggplot2::element_blank(),
+          axis.ticks.length = grid::unit(0, "null")
+        )
+      
+      # since the tick marks are removed on the marginal plot, we need to add
+      # space for them so that the marginal plot will align with the main plot
+      if (is.null(p$theme$axis.ticks.length)) {
+        marginUnit <- "null"
+        marginLength <- 0
+      } else {
+        marginUnit <- attr(p$theme$axis.ticks.length, "unit")
+        marginLength <- as.numeric(p$theme$axis.ticks.length, "unit")
+      }
+      if (margin == "x") {
+        marginal <- marginal + 
+          ggplot2::theme(
+            axis.title.x = ggplot2::element_blank(),
+            axis.text.x = ggplot2::element_blank(),
+            plot.margin = grid::unit(c(0, 0, 0, marginLength), marginUnit)
+          )
+      } else {
+        marginal <- marginal + 
+          ggplot2::theme(
+            axis.title.y = ggplot2::element_blank(),
+            axis.text.y = ggplot2::element_blank(),
+            plot.margin = grid::unit(c(0, 0, marginLength, 0), marginUnit)
+          )
+      }
+    }
+    # if this is the old ggplot2 version, things are simpler
+    else {
+      marginal <- marginal +
+        ggplot2::theme(
+          text = ggplot2::element_text(size = p$theme$text$size, color = "transparent"),
+          line = ggplot2::element_blank(),
+          panel.background = ggplot2::element_blank(),
+          axis.text = ggplot2::element_text(color = "transparent")
+        )
+
+      if (margin == "x") {
+        marginal <- marginal + 
+          ggplot2::theme(
+            axis.title.x = ggplot2::element_blank(),
+            axis.text.x = ggplot2::element_blank(),
+            plot.margin = grid::unit(c(0, 0, -1, 0), "lines")
+          )
+      } else {
+        marginal <- marginal + 
+          ggplot2::theme(
+            axis.title.y = ggplot2::element_blank(),
+            axis.text.y = ggplot2::element_blank(),
+            plot.margin = grid::unit(c(0, 0, 0, -1), "lines")
+          )
+      }
+    }
+    
+    marginal
+  }
+  
   # Copy the scale transformation from the original plot (reverse/log/limits/etc)
   # We have to do a bit of a trick on the marginal plot that's flipped by
   # taking the original x/y scale and manually changing it to the other axis
-  get_scale <- function(margin) {
+  getScale <- function(margin) {
     if (margin == "x") {
       if (type == "boxplot") {
         scale <- pb$panel$x_scales[[1]]
@@ -218,7 +325,7 @@ ggMarginal <- function(p, data, x, y, type = c("density", "histogram", "boxplot"
   # Get the axis range of the x or y axis of the given ggplot build object
   # This is needed so that if the range of the plot is manually changed, the
   # marginal plots will use the same range
-  get_limits <- function(pb, margin) {
+  getLimits <- function(pb, margin) {
     if (margin == "x") {
       scales <- pb$panel$x_scales[[1]]
     } else if (margin == "y") {
@@ -246,18 +353,11 @@ ggMarginal <- function(p, data, x, y, type = c("density", "histogram", "boxplot"
   # - Use the same axis titles as the main plot, to ensure the same space is taken
   # - Use the same axis range as the main plot
   if (margins != "y") {
-    top <-
-      marginPlot("x") + 
-      ggplot2::theme(
-        text = ggplot2::element_text(size = textsize, color = "transparent"),
-        line = ggplot2::element_blank(),
-        panel.background = ggplot2::element_blank(),
-        axis.text = ggplot2::element_text(color = "transparent"),
-        axis.title.x = ggplot2::element_blank(),
-        axis.text.x = ggplot2::element_blank(),
-        plot.margin = grid::unit(c(0, 0, -1, 0), "lines")) +
+    top <- marginPlot("x")
+    top <- addMainTheme(top, "x")
+    top <- top +
       ggplot2::ylab(p$labels$y) +
-      get_scale("x")
+      getScale("x")
     
     # Add the longest y axis label to the top plot and ensure it's at a y value
     # that is on the plot (this is why I build the top plot, to know the y values)
@@ -267,46 +367,26 @@ ggMarginal <- function(p, data, x, y, type = c("density", "histogram", "boxplot"
     if (type == "boxplot") {
       top <-
         top +
-        ggplot2::scale_x_continuous(breaks = mean(get_limits(pbTop, "x")),
+        ggplot2::scale_x_continuous(breaks = mean(getLimits(pbTop, "x")),
                                     labels = ylabel)      
     } else {
       top <-
         top +
-        ggplot2::scale_y_continuous(breaks = mean(get_limits(pbTop, "y")),
+        ggplot2::scale_y_continuous(breaks = mean(getLimits(pbTop, "y")),
                                     labels = ylabel)      
-    }
-
-    
-    # If we are showing a marginal plot above the main plot, then transfer the
-    # plot title to be above the marginal plot
-    # TODO(daattali) This doesn't quite work right because the title is taking
-    # a lot of vertical real estate from the short marginal plot. It needs to
-    # go above the marginal plot without affecting its height.
-    if (FALSE) {
-      if (!is.null(pb$plot$labels$title)) {
-        top <- top + ggplot2::ggtitle(pb$plot$labels$title)
-        p <- p + ggplot2::ggtitle("")
-      }
     }
   }
 
   # Create the vertical margin plot
   if (margins != "x") {
-    right <-
-      marginPlot("y") + 
-      ggplot2::theme(
-        text = ggplot2::element_text(size = textsize, color = "transparent"),
-        line = ggplot2::element_blank(),
-        panel.background = ggplot2::element_blank(),
-        axis.text = ggplot2::element_text(color = "transparent"),
-        axis.title.y = ggplot2::element_blank(),
-        axis.text.y = ggplot2::element_blank(),
-        plot.margin = grid::unit(c(0, 0, 0, -1), "lines")) +
+    right <- marginPlot("y")
+    right <- addMainTheme(right, "y")
+    right <- right +
       ggplot2::ylab(p$labels$x) +
-      get_scale("y") +
+      getScale("y") +
       ggplot2::ggtitle(p$labels$title)
   }
-
+  
   # Build a 2x2 or 2x1 grid to arrange the plots  
   ncol <- 2
   nrow <- 2
@@ -349,6 +429,7 @@ ggMarginal <- function(p, data, x, y, type = c("density", "histogram", "boxplot"
   
   # Build the grid of plots
   suppressMessages(
+    # use suppressMessages to ignore message about adding multiple scales
     plot <- do.call(gridExtra::arrangeGrob, gridArgs)
   )
   class(plot) <- c("ggExtraPlot", class(plot))